backend/app/infrastructure/auth/password.py
"""Funções de hash e verificação de senha usando bcrypt diretamente."""
import bcrypt


def hash_password(password: str) -> str:
    """Gera hash bcrypt de uma senha."""
    password_bytes = password.encode("utf-8")
    salt = bcrypt.gensalt()
    hashed = bcrypt.hashpw(password_bytes, salt)
    return hashed.decode("utf-8")


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verifica se uma senha corresponde ao hash."""
    return bcrypt.checkpw(
        plain_password.encode("utf-8"),
        hashed_password.encode("utf-8"),
    )
